Basic Expansion on Networking
I expanded on the network implementation and it seems that all is required right now is the implementation of networking components into the object that require them.

I also added the code required for the servers, which is split into a TCP and UDP server and should seamlessly integrate with the game. However, all of this is not tested in practice, so that's why I branched this off into it;s own repo.

Read NETWORKING_INTEGRATION_COMPLETE.md for further details.